backend/assistant/requirements.txt
groq>=0.9.0
psycopg2-binary